+++ resources/views/ManufactureManage/Processinfos/index.blade.php
@@ -1,11 +1,11 @@
 @extends('navbarerp')
 @section('title', '工艺单信息')
 @section('main')
-
+    @can('new_processinfo')
     <div class="panel-heading">
         <a href="/ManufactureManage/Processinfos/create" class="btn btn-sm btn-success">新建(New)</a>
     </div>
-
+    @endcan
     <div class="panel-body">
         {!! Form::open(['url' => '/ManufactureManage/Processinfos/search', 'class' => 'pull-right form-inline', 'id' => 'frmSearch']) !!}
         <div class="form-group-sm">
@@ -65,11 +65,15 @@
                                 {{ $processinfo->created_at }}
                             </td>
                             <td>
-                                <a href="{{ URL::to('/ManufactureManage/Processinfos/'.$processinfo->id.'/copy') }}" class="btn btn-warning btn-sm pull-left">Copy</a>
-                                <a href="{{ URL::to('/ManufactureManage/Processinfos/'.$processinfo->id.'/edit') }}" class="btn btn-success btn-sm pull-left">Edit</a>
-                                {!! Form::open(array('route' => array('Processinfos.destroy', $processinfo->id), 'method' => 'delete', 'onsubmit' => 'return confirm("确定删除此记录(Confirm to delete this record)?");')) !!}
-                                    {!! Form::submit('Del', ['class' => 'btn btn-danger btn-sm pull-left']) !!}
-                                {!! Form::close() !!}
+                                @can('edit_peocessinfo')
+                                    <a href="{{ URL::to('/ManufactureManage/Processinfos/'.$processinfo->id.'/copy') }}" class="btn btn-warning btn-sm pull-left">Copy</a>
+                                @endcan
+                                @can('del_processinfo')
+                                    <a href="{{ URL::to('/ManufactureManage/Processinfos/'.$processinfo->id.'/edit') }}" class="btn btn-success btn-sm pull-left">Edit</a>
+                                    {!! Form::open(array('route' => array('Processinfos.destroy', $processinfo->id), 'method' => 'delete', 'onsubmit' => 'return confirm("确定删除此记录(Confirm to delete this record)?");')) !!}
+                                        {!! Form::submit('Del', ['class' => 'btn btn-danger btn-sm pull-left']) !!}
+                                    {!! Form::close() !!}
+                                @endcan
                                 {{--<a href="{{ URL::to('/development/fabricdischarges/'.$fabricdischarge->id . '/export') }}" class="btn btn-success btn-sm pull-left">导出</a>--}}
                             </td>
                         </tr>
